refactor(core): inline memory and runtime modules
