+++ Cargo.toml
@@ -6,14 +6,6 @@ description = "Core of the AEI Framework for dynamic neural networks."
 license = "MPL-2.0"
 publish = false
 
-[workspace]
-members = [
-    "crates/memory",
-    "crates/runtime",
-    "crates/utils",
-]
-resolver = "2"
-
 [dependencies]
 uuid = { version = "1.8", features = ["v4", "serde"] }
 serde = { version = "1.0", features = ["derive"] }
@@ -22,9 +14,9 @@ log = "0.4"
 rand = "0.8"
 rand_distr = "0.4"
 chrono = { version = "0.4", features = ["serde"] }
+thiserror = "1.0"
+crossbeam-channel = "0.5"
 
 [dev-dependencies]
 env_logger = "0.10"
 rand_chacha = "0.3"
-aei-memory = { path = "crates/memory" }
-aei-runtime = { path = "crates/runtime" }
