--- crates/runtime/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "aei-runtime"
-version = "0.1.0"
-edition = "2021"
-description = "Agent runtime responsible for scheduling and orchestration."
-license = "MPL-2.0"
-
-[dependencies]
-crossbeam-channel = "0.5"
