--- crates/runtime/src/scheduler/mod.rs
@@ -1,5 +1,0 @@
-//! Task scheduling utilities.
-
-mod scheduler;
-
-pub use scheduler::{InMemoryScheduler, Scheduler};
