--- crates/utils/src/lib.rs
@@ -1,2 +1,0 @@
-//! Shared utility functions and types used across the
-//! Autonomous & Evolutive Intelligence Framework (AEIF).
