+++ docs/fr/scheduler.md
@@ -3,7 +3,7 @@
 Le module Scheduler enregistre des tâches exécutées à intervalles fixes. Les tâches sont lancées lorsque `tick()` est appelé.
 
 ```rust
-use aei_runtime::scheduler::{InMemoryScheduler, Scheduler};
+use aei_framework::core::scheduler::{InMemoryScheduler, Scheduler};
 use std::time::Duration;
 
 let mut sched = InMemoryScheduler::new();
