+++ src/core/memory/compactor.rs
