+++ src/core/memory/events.rs
