+++ src/core/memory/index.rs
