+++ src/core/memory/mod.rs
