+++ src/core/memory/retention.rs
