+++ src/core/mod.rs
@@ -1,0 +1,5 @@
+//! Core modules with generic utilities.
+
+pub mod event_bus;
+pub mod memory;
+pub mod scheduler;
