+++ src/lib.rs
@@ -2,6 +2,7 @@
 //! modular neural networks.
 
 pub mod application;
+pub mod core;
 pub mod domain;
 pub mod infrastructure;
